Fire.js created
